[Emacs] - Configured Emacs to run in Daemon mode (optional)
1) Start emacs using emacs --daemon
2) Start emacs with the alias 'e' for a very fast startup

The last line in .emacs must be '(setq debug-on-error nil)'
otherwise Lisp debugging mode is turned on and we are thrown
in to a Lisp debugging frame. This happens every time a function
returns an error, that is it happens when you go the next error
and there is no more errors to go to.